wip: collect css information in worker
Even though this plugin is smarter than using postcss with globs, it's still a pretty significant contributor to Vite slowness (in our large app, it adds 5-10 seconds to the first page load, depending on how big the first page load is)

if there are no transformations being done in the plugin, it could make sense to do the parsing in a seperate worker so that it doesn't slow down the main Vite transformation pipeline.

This just proves out the idea, and isn't ready to be merged or anything. See "TODO" comments for a number of things that would be needed if we decide this is worth doing.

You can see the benefit by running the make-files.sh script, then doing `pnpm dev` in the playground with/without the "worker" option in the vite plugin.